# api/swagger.yaml
definitions:
  models.DefaultErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Login:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      login:
        type: string
      password:
        type: string
      user_id:
        type: integer
    type: object
info:
  contact: {}
paths:
  /user/balance:
    get:
      consumes:
      - application/json
      description: get user balance
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
      summary: balance
      tags:
      - balance
  /user/balance/withdraw:
    post:
      consumes:
      - application/json
      description: withdraw balance
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
      summary: withdraw
      tags:
      - balance
  /user/login:
    post:
      consumes:
      - application/json
      description: user login
      parameters:
      - description: login
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: login
      tags:
      - auth
  /user/orders:
    get:
      consumes:
      - application/json
      description: order get
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.DefaultErrorResponse'
        "500":
          description: Internal Server Error
      summary: get
      tags:
      - orders
    post:
      consumes:
      - text/plain
      description: order create
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DefaultErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.DefaultErrorResponse'
        "409":
          description: Conflict
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: create
      tags:
      - orders
  /user/register:
    post:
      consumes:
      - application/json
      description: user registration
      parameters:
      - description: user account
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DefaultErrorResponse'
        "404":
          description: Not Found
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.DefaultErrorResponse'
        "500":
          description: Internal Server Error
      summary: register
      tags:
      - auth
  /user/withdrawals:
    get:
      consumes:
      - application/json
      description: user withdrawals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "204":
          description: No Content
      summary: withdrawals
      tags:
      - withdrawals
swagger: "2.0"