# internal/app/gophermart/controllers/orders.go
package controllers

import (
	"github.com/IgorChicherin/gophermart/internal/app/gophermart/middlewares"
	"github.com/IgorChicherin/gophermart/internal/app/gophermart/usecases"
	"github.com/IgorChicherin/gophermart/internal/pkg/accrual"
	"github.com/ShiraazMoollatjie/goluhn"
	"github.com/gin-gonic/gin"
	log "github.com/sirupsen/logrus"
	"net/http"
)

type OrdersController struct {
	OrderUseCase   usecases.OrderUseCase
	AccrualService accrual.AccrualService
}

func (oc OrdersController) Route(api *gin.RouterGroup) {
	middleware := middlewares.AuthMiddleware(oc.OrderUseCase.GetUserRepository())
	orders := api.Group("/user").Use(middleware)
	{
		orders.POST("/orders", oc.orderCreate)
		orders.GET("/orders", oc.orderGet)
	}
}

// @BasePath /api
// login godoc
// @Summary create
// @Schemes
// @Description order create
// @Tags orders
// @Accept plain
// @Produce json
// @Success 200,202
// @Failure 409,422,500
// @Failure 400,401 {object} models.DefaultErrorResponse
// @Router /user/orders [post]
func (oc OrdersController) orderCreate(c *gin.Context) {
	b, err := c.GetRawData()

	if err != nil {
		log.WithFields(log.Fields{"func": "orderCreate"}).Errorln(err)
		c.AbortWithStatus(http.StatusInternalServerError)
		return
	}

	orderNr := string(b)

	token, err := c.Cookie("token")
	userRepo := oc.OrderUseCase.GetUserRepository()

	if err != nil {
		log.WithFields(log.Fields{"func": "orderCreate"}).Errorln(err)
		c.AbortWithStatusJSON(http.StatusUnauthorized, map[string]string{"error": "unauthorized user"})
		return
	}

	login, _, err := userRepo.DecodeToken(token)

	if err != nil {
		log.WithFields(log.Fields{"func": "orderCreate"}).Errorln(err)
		c.AbortWithStatus(http.StatusInternalServerError)
		return
	}

	err = goluhn.Validate(orderNr)
	if err != nil {
		log.WithFields(log.Fields{"func": "orderCreate"}).Errorln(err)
		c.AbortWithStatus(http.StatusUnprocessableEntity)
		return
	}

	orderRepo := oc.OrderUseCase.GetOrderRepository()
	hasOrder, err := orderRepo.HasOrder(orderNr)

	if err != nil {
		log.WithFields(log.Fields{"func": "orderCreate"}).Errorln(err)
		c.AbortWithStatus(http.StatusInternalServerError)
		return
	}

	if hasOrder {
		order, err := orderRepo.GetOrder(orderNr)

		if err != nil {
			log.WithFields(log.Fields{"func": "orderCreate"}).Errorln(err)
			c.AbortWithStatus(http.StatusInternalServerError)
			return
		}

		user, err := userRepo.GetUser(login)

		if err != nil {
			log.WithFields(log.Fields{"func": "orderCreate"}).Errorln(err)
			c.AbortWithStatus(http.StatusInternalServerError)
		}

		if order.UserID != user.UserID {
			c.AbortWithStatus(http.StatusConflict)
			return
		}

		c.AbortWithStatus(http.StatusOK)
		return
	}

	_, err = oc.OrderUseCase.CreateOrder(login, orderNr)

	if err != nil {
		log.WithFields(log.Fields{"func": "orderCreate"}).Errorln(err)
		c.AbortWithStatus(http.StatusInternalServerError)
		return
	}

	go func() {
		oc.AccrualService.ProcessOrder(orderNr)
	}()
	c.Status(http.StatusAccepted)
}

// @BasePath /api
// login godoc
// @Summary get
// @Schemes
// @Description order get
// @Tags orders
// @Accept json
// @Produce json
// @Success 200 {json} []models.OrderListItem
// @Success 204
// @Failure 401 {object} models.DefaultErrorResponse
// @Failure 500
// @Router /user/orders [get]
func (oc OrdersController) orderGet(c *gin.Context) {
	token, err := c.Cookie("token")
	userRepo := oc.OrderUseCase.GetUserRepository()

	if err != nil {
		log.WithFields(log.Fields{"func": "orderGet"}).Errorln(err)
		c.AbortWithStatusJSON(http.StatusUnauthorized, map[string]string{"error": "unauthorized user"})
		return
	}

	login, _, err := userRepo.DecodeToken(token)

	if err != nil {
		log.WithFields(log.Fields{"func": "orderGet"}).Errorln(err)
		c.AbortWithStatus(http.StatusInternalServerError)
		return
	}

	ordersList, err := oc.OrderUseCase.GetOrdersList(login)
	if err != nil {
		log.WithFields(log.Fields{"func": "orderGet"}).Errorln(err)
		c.AbortWithStatus(http.StatusInternalServerError)
		return
	}

	if len(ordersList) == 0 {
		c.AbortWithStatus(http.StatusNoContent)
		return
	}

	c.AbortWithStatusJSON(http.StatusOK, ordersList)
}